refactor: move third party code to a separate subfolder
