+++ fs2/attn/attention.py
